Add session/auth wiring, admin protection, input validation, Global.h
Bug fixes :
Parser: read user permissions as int32_t. They were cast to uint8_t,
which cut the 12-bit mask to 8 bits and turned -1 (all) into 255.
LoginScreen: return nullptr after too many failed attempts (the
function used to end without a return).
Authorizer: choosing Manage Users now sets the ManageUsers bit
(before, only the user-management sub-permissions were set).
UserRepository::AddUser: duplicate check now compares the username
only (it also compared the password hash before).

Permissions and admin :
Admin bypasses every permission check in Authorizer::HasAccess.
Authorizer::IsAdmin is public and compares the stored hash with
ADMIN_HASH directly (it used to bcrypt-verify the hash string).
Selecting all 12 permission bits stores -1 (full access);
IsPermissionsChanged treats -1 and 4095 as equal.
Granting user-management permissions first checks that the granting
user has something to give, then asks until at least one is chosen.
Adds two unused _SetPermission helpers to Authorizer.

User management :
The Admin account can no longer be deleted or updated
(new UserIsAdmin state and UserIsAdminMsg).
Delete/update now receive the current user. If a user deletes their
own account, or changes their own username, password or permissions,
the session user is cleared and a note asks them to log out and sign
in again (new SuccessfulSelfDelete / SuccessfulSelfUpdate states and
SelfEditMsg).
The delete/update user screens still ask for a password but no longer
check it: Find/IsExists/IsExistsInFile lost their password parameter.
Field-by-field update: ClientRepository::UpdateExistingClient (first
name, last name, email, phone, PIN) and UserRepository::
UpdateExistingUser (also username and password) ask (change X?) per
field and keep the rest. UpdateScreen now uses the client version.
UserRepository::_UpdateObject no longer saves each object before the
whole list is saved.
Client and user cards and lists no longer print the PIN/password hash.

Input validation :
 Validator::IsInvalid rejects input containing # or / (characters
of the #//# delimiter) so records cannot be corrupted.
Validator::ReadNames and ReadEmails re-prompt while input is invalid.
 Applied to: account number when adding a client; first name, last
name and email when updating a client; username, first name, last
name and email when adding or updating a user.
Phone prompt reworded; pincode entry now prints a heading.

Session, UI and main loop :
 Session: GetUser() is non-const only; adds isActiveState() and
isEndedState(); isEnded() now returns !m_isActive (its old logic is
now isEndedState()).
 UI: adds Start(User&) and PerformMenu(User&) overloads.
Screen: adds UserIsAdminMsg and SelfEditMsg.
BankOOP.cpp: the main loop runs on Session::isActiveState(). Logout
now ends the program instead of returning to the login screen.

Global.h :
 New file with ADMIN_USERNAME, ADMIN_HASH, FullPermissions (-1),
FullPermisssionsInPositive (4095), FullPermssionsCount (12, unused),
CLIENTS_FILE, USERS_FILE and DELIMITER.
Removed from FileHandler.h (file names, delimiter) and Authorizer.h
(admin username/hash). FileHandler.cpp, Authorizer.h, Validator.h and
UserRepository.h include Global.h.

+++ Parser.cpp
@@ -34,6 +34,6 @@ User Parser::LineToUser(std::string line) {
 		throw std::runtime_error("Malformed line: expected 7 fields, got " + std::to_string(Tokens.size()));
 
 
-	return User(std::move(Tokens[0]), std::move(Tokens[1]), std::move(Tokens[2]), std::move(Tokens[3]), std::move(Tokens[4]), std::move(Tokens[5]), static_cast<uint8_t>(stoi(Tokens[6])), User::ObjectMode::ExistingMode);
+	return User(std::move(Tokens[0]), std::move(Tokens[1]), std::move(Tokens[2]), std::move(Tokens[3]), std::move(Tokens[4]), std::move(Tokens[5]), static_cast<int32_t>(stoi(Tokens[6])), User::ObjectMode::ExistingMode);
 
 }
